# Add performance testing setup (close #37)