--- a/performance/run.cpp
+++ b/performance/run.cpp
@@ -9,6 +9,8 @@
 #include "mute_emitter.hpp"
 #include "mock_client_session.hpp"
 
+void clear_storage(const string &db_name);
+
 void track_events() {
   Tracker *tracker = Tracker::instance();
 
@@ -74,6 +76,7 @@ double run_mocked_emitter_and_mocked_session(const string &db_name) {
 double run_mocked_emitter_and_real_session(const string &db_name) {
   MockEmitter emitter(db_name);
   ClientSession client_session(db_name, 5000, 5000, 2500);
+  clear_storage(db_name);
   double time = run(emitter, client_session);
   Storage::close();
   return time;
@@ -82,6 +85,7 @@ double run_mocked_emitter_and_real_session(const string &db_name) {
 double run_mute_emitter_and_mocked_session(const string &db_name) {
   MuteEmitter emitter(db_name);
   MockClientSession client_session(db_name);
+  clear_storage(db_name);
   double time = run(emitter, client_session);
   Storage::close();
   return time;
@@ -90,7 +94,14 @@ double run_mute_emitter_and_mocked_session(const string &db_name) {
 double run_mute_emitter_and_real_session(const string &db_name) {
   MuteEmitter emitter(db_name);
   ClientSession client_session(db_name, 5000, 5000, 2500);
+  clear_storage(db_name);
   double time = run(emitter, client_session);
   Storage::close();
   return time;
+}
+
+void clear_storage(const string &db_name) {
+  Storage *storage = Storage::init(db_name);
+  storage->delete_all_event_rows();
+  storage->delete_all_session_rows();
 }