@matus-tomlein 2022-02-14
@mscwilson Hey Miranda, I thought I'd tag you here in case this is interesting for you since you were working on similar performance testing tasks. No need to review the C++ code, probably more interesting is the general approach and procedure for performance testing. If you have any suggestions or would like to chat about it, let me know.