@mscwilson 2022-02-15
For the Python comparison table, can you add a column for `n`? Would give more context to the mean and other values.