@matus-tomlein 2022-02-16
Thanks for the comments, both!

Miranda, I changed the Python script to also show the number of operations and threads.

Alex, I don't have a clear idea yet how to integrate with CI. It would be ideal to run the tests on every tagged release, but not sure how to store the results of the test. One option would be to add it as asset to the release as Paul suggested yesterday. It requires some more thought, let's first see how useful the tests are as we change stuff in the tracker.